pyproject.toml
[project]
name = "coinbasekafkaservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "coinbase-advanced-py",
    "python-dotenv",
]
